tests/wpt/web-platform-tests/dom/nodes/CharacterData-insertData.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>CharacterData.insertData</title>
<link rel=help href="https://dom.spec.whatwg.org/#dom-characterdata-insertdata">
<link rel=help href="https://dom.spec.whatwg.org/#dom-characterdata-data">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
function testNode(create, type) {
  test(function() {
    var node = create()
    assert_equals(node.data, "test")

    assert_throws("INDEX_SIZE_ERR", function() { node.insertData(5, "x") })
    assert_throws("INDEX_SIZE_ERR", function() { node.insertData(5, "") })
  }, type + ".insertData() out of bounds")

  test(function() {
    var node = create()
    assert_equals(node.data, "test")

    node.insertData(2, "X")
    assert_equals(node.data, "teXst")
  }, type + ".insertData() in the middle")

  test(function() {
    var node = create()
    assert_equals(node.data, "test")

    node.insertData(4, "ing")
    assert_equals(node.data, "testing")
  }, type + ".insertData() at the end")
}

testNode(function() { return document.createTextNode("test") }, "Text")
testNode(function() { return document.createComment("test") }, "Comment")
</script>
